+++ themes/box-of-scraps/archive.php
@@ -14,14 +14,13 @@
 
 <?php
 /** This action is documented in includes/Linchpin/truss-hooks.php */
-do_action( 'truss_content_before' );
+do_action( 'boxofscraps_content_before' );
 ?>
 
 	<div class="container py-4">
 		<?php if ( have_posts() ) : ?>
 			<?php
-			/** This action is documented in includes/Linchpin/truss-hooks.php */
-			do_action( 'truss_loop_before' );
+			do_action( 'boxofscraps_loop_before' );
 			?>
 
 			<div class="columns is-multiline">
@@ -33,8 +32,7 @@ do_action( 'truss_content_before' );
 			</div>
 
 			<?php
-			/** This action is documented in includes/Linchpin/truss-hooks.php */
-			do_action( 'truss_loop_after' );
+			do_action( 'boxofscraps_loop_after' );
 			?>
 
 		<?php else : ?>
@@ -45,8 +43,7 @@ do_action( 'truss_content_before' );
 <?php get_template_part( 'partials/pagination' ); ?>
 
 <?php
-/** This action is documented in includes/Linchpin/truss-hooks.php */
-do_action( 'truss_content_after' );
+do_action( 'boxofscraps_content_after' );
 ?>
 
 <?php get_footer();