+++ themes/box-of-scraps/single.php
@@ -1,27 +1,24 @@
 <?php
 /**
- * Page Template
+ * Single Post Template
  *
  * @since      1.0.0
  *
  * @package    BoxOfScraps
  * @subpackage Templates
  */
-
 ?>
 
 <?php get_header(); ?>
 
 <?php
-/** This action is documented in includes/Linchpin/truss-hooks.php */
-do_action( 'truss_content_before' );
+do_action( 'boxofscraps_content_before' );
 ?>
 
 <div class="container py-4">
 	<?php if ( have_posts() ) : ?>
 		<?php
-		/** This action is documented in includes/Linchpin/truss-hooks.php */
-		do_action( 'truss_loop_before' );
+		do_action( 'boxofscraps_loop_before' );
 		?>
 
 		<div class="columns">
@@ -38,8 +35,7 @@ do_action( 'truss_content_before' );
 		</div>
 
 		<?php
-		/** This action is documented in includes/Linchpin/truss-hooks.php */
-		do_action( 'truss_loop_after' );
+		do_action( 'boxofscraps_loop_after' );
 		?>
 
 		<?php
@@ -55,8 +51,8 @@ do_action( 'truss_content_before' );
 </div>
 
 <?php
-/** This action is documented in includes/Linchpin/truss-hooks.php */
-do_action( 'truss_content_after' );
+do_action( 'boxofscraps_content_after' );
 ?>
 
-<?php get_footer();
+<?php
+get_footer();