+++ themes/box-of-scraps/CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.6] - 04-05-2021
+
+### Added
+- Added utility class to pull in SVG Icons
+- New partial for the hamburger icon
+- Added manifest.json for SEO
+
+### Updated
+- Updated custom Walker_Nav_Menu name to include "Bulma" Bulma_Walker_Nav_Menu
+- Updated comments template
+
+### Removed
+- Removed any references to truss
+
 ## [1.0.0] - 03-25-2021
 ### Added
 - Initial Release